requirements.txt
fastapi>=0.78.0
uvicorn>=0.18.0
pymongo>=4.0.0
pandas>=1.4.0
scikit-learn>=1.0.0
openai>=0.27.0
python-dotenv>=0.20.0